Functions

Managing sets of taxa

Monstera.TaxonSet — Type

TaxonSet([taxon_names])
TaxonSet(n::Int)

An ordered collection of taxa with unique names.

TaxonSets can be created empty, or from a vector of name strings, or by giving a number n, which creates taxa successively named "1" through "n".

Base.get — Method

get(taxonset, name)

Get the taxon ID that matches a name in a taxon set, or return 0 if there is no match.

Monstera.key — Method

key(taxonset, ids...)

Get the names of the taxa with the given ids in the taxon set.

Base.haskey — Function

haskey(collection, key) -> Bool

Determine whether a collection has a mapping for a given key.

Examples

julia> D = Dict('a'=>2, 'b'=>3)
Dict{Char, Int64} with 2 entries:
  'a' => 2
  'b' => 3

julia> haskey(D, 'a')
true

julia> haskey(D, 'c')
false

haskey(sc,k)

Returns true if key k is present for SortedDict, SortedMultiDict or SortedSet sc. For SortedSet, haskey(sc,k) is a synonym for in(k,sc). For SortedDict and SortedMultiDict, haskey(sc,k) is equivalent to in(k,keys(sc)). Time: O(c log n)

haskey(sc,k)

Returns true if key k is present for SortedDict, SortedMultiDict or SortedSet sc. For SortedSet, haskey(sc,k) is a synonym for in(k,sc). For SortedDict and SortedMultiDict, haskey(sc,k) is equivalent to in(k,keys(sc)). Time: O(c log n)

haskey(sc,k)

Returns true if key k is present for SortedDict, SortedMultiDict or SortedSet sc. For SortedSet, haskey(sc,k) is a synonym for in(k,sc). For SortedDict and SortedMultiDict, haskey(sc,k) is equivalent to in(k,keys(sc)). Time: O(c log n)

haskey(collection, key) -> Bool

Determine whether a collection has a mapping for a given key.

Examples

julia> D = RobinDict('a'=>2, 'b'=>3)
RobinDict{Char,Int64} with 2 entries:
  'a' => 2
  'b' => 3

julia> haskey(D, 'a')
true

julia> haskey(D, 'c')
false

haskey(collection, key) -> Bool

Determine whether a collection has a mapping for a given key.

Examples

julia> D = OrderedRobinDict('a'=>2, 'b'=>3)
OrderedRobinDict{Char,Int64} with 2 entries:
  'a' => 2
  'b' => 3

julia> haskey(D, 'a')
true

julia> haskey(D, 'c')
false

haskey(collection, key) -> Bool

Determine whether a collection has a mapping for a given key.

Examples

julia> D = SwissDict('a'=>2, 'b'=>3)
SwissDict{Char,Int64} with 2 entries:
  'a' => 2
  'b' => 3

julia> haskey(D, 'a')
true

julia> haskey(D, 'c')
false

haskey(tree, key)

Returns true if key is present in the tree, else returns false.

haskey(taxonset, name)

Check whether a taxon set contains a name.

Base.keys — Method

keys(taxonset)

Return the list of names of taxa in a taxon set.

Monstera.add! — Method

add!(taxonset, names...)

Add name(s) to a taxon set and return its id.

Names already in the taxon set will be ignored (skipped or return 0).

Creating trees

Monstera.clone — Function

clone(tree)

Create a copy of a tree.

The copied tree has the same topology and properites, but its contents are fully independent. Only the taxon set is shared between the original tree and the clone.

Monstera.randtree — Function

randtree(T, taxonset, rooted=true)
randtree(T, n, rooted=true)

Create a random binary tree of type T with the given taxon set, or the given number n of outer nodes.

If rooted is false, the tree will be unrooted, i.e. it will have a basal trichotomy (this requires at least 3 outer nodes).

Random.seed!(1)
textplot(randtree(RoundaboutTree, 5, false))

# output

┌────○(2)
│
│    ┌───────○(5)
○(1)─○(3)
│    └───────○(6)
│
│    ┌───────○(7)
└────○(4)
     └───────○(8)

Reading and writing Newick trees

Monstera.Newick.parse — Function

Newick.parse(T, string, settings)

Parse a string in Newick format as a tree of type T.

Monstera.Newick.read — Function

Newick.read(file, sink)
Newick.read(io, sink)

Read trees from a file in Newick format.

The sink argument specifies the concrete type that will be used to represent the trees. It can correspond to an AbstractTree subtype or a vector thereof. If a tree type is given, only the first Newick string in the file will be read. If a vector of a tree type is given, all the Newick strings in the file will be read.

Monstera.Newick.string — Function

Newick.string(tree, <keyword arguments or writer settings object>)
Newick.string(node, <keyword arguments or writer settings object>)
Newick.string(branch => node, <keyword arguments or writer settings object>)

Return a string with the Newick representation of a tree.

The string format can be optionally customised by passing a WriterSettings object, or with the keyword arguments below.

See also Newick.read, Newick.WriterSettings

Arguments

brlength::Bool=true: whether to write branch lengths

label_quote::Bool=false: wheteher to write double quote marks around node labels

brlength_digits::Int=308: branch lengths will be rounded to this number

skip_nanlength::Bool=false: whether to write the length of a branch when it is NaN

Monstera.Newick.write — Function

Newick.write(<file name or io>, tree, <keyword arguments or writer settings object>)
Newick.write(<file name or io>, node, <keyword arguments or writer settings object>)
Newick.write(<file name or io>, branch => node, <keyword arguments or writer settings object>)

Write the Newick representation of a tree to an IOBuffer or a file.

The Newick string format can be optionally customised by passing a NewickWriterSettings object, or with the keyword arguments below.

See also Newick.string, Newick.WriterSettings

Arguments

brlength::Bool=true: whether to write branch lengths

label_quote::Bool=false: wheteher to write double quote marks around node labels

brlength_digits::Int=308: branch lengths will be rounded to thisnumber

skip_nanlength::Bool=false: whether to write the length of a branch when it is NaN

Writing to file

append=false: whether to append to the file

lock=true: lock operations for safe multi-threaded access.

Monstera.Newick.WriterSettings — Type

Newick.WriterSettings(args)

Settings for writing Newick strings.

Arguments

brlength::Bool=true: whether to write branch lengths

label_quote::Bool=false: wheteher to write double quote marks around node labels

brlength_digits::Int=308: branch lengths will be rounded to this number of digits

skip_nanlength::Bool=true: whether to write the length of a branch when it is NaN

Monstera.Newick.ReaderSettings — Type

Newick.ReaderSettings(taxonset = TaxonSet(); taxafrom=:outer)

Settings for reading Newick strings.

Node labels can be interpreted as taxon names. Taxon names will be mapped to the corresponding taxon in the taxon set, or added to the taxon set if they were not already present.

The taxafrom keyword argument is used to customise which node labels will be interpeted as taxon names, with any of the following values:

:outer (default) only outer nodes

:inner only inner nodes

:all all nodes

:none no nodes

Branches and nodes

Monstera.neighbours — Function

neighbours(node) -> iterator

Return an iterator of the branches and nodes that are connected to a given node.

See also children.

Monstera.children — Function

children(branch => node) -> iterator
children(node) -> iterator

Return an iterator of the branches and nodes that are children of a node.

See also neighbours.

Monstera.isinner — Function

isinner(node) -> Bool

Check whether a node is inner (a.k.a. "internal").

Inner nodes have a degree of 2 or greater.

See also isouter.

Monstera.isouter — Function

isouter(node) -> Bool

Check whether a node is outer (a.k.a. "external", "tip", or "leaf").

Outer nodes always have a degree of 1. This can include the tree root and isolated nodes.

See also isinner.

Monstera.nodes_flanking — Function

nodes_flanking(branch)

Return a tuple with the nodes connected to a branch.

When nodes are missing from one side of the branch, nothing will take their place in the tuple.

See also branch_between.

Monstera.branch_between — Function

branch_between(node1, node2)

Return the branch that connects node1 to node2.

See also nodes_flanking.

Monstera.hasparent — Function

hasparent(node)

Check whether a node has a parent.

Base.parent — Function

parent(A)

Return the underlying "parent array”. This parent array of objects of types SubArray, ReshapedArray or LinearAlgebra.Transpose is what was passed as an argument to view, reshape, transpose, etc. during object creation. If the input is not a wrapped object, return the input itself. If the input is wrapped multiple times, only the outermost wrapper will be removed.

Examples

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> V = view(A, 1:2, :)
2×2 view(::Matrix{Int64}, 1:2, :) with eltype Int64:
 1  2
 3  4

julia> parent(V)
2×2 Matrix{Int64}:
 1  2
 3  4

parent(node)
parent(branch => node)

Return the parent of a node.

Throws an error if the node has no parent.

Monstera.link! — Function

link!(node1, branch, node2)

Connect node1 to node2 through a branch.

Linking order is left-to-right, so that node1 is connected on the left side of the branch and node2 is connected on the right side of the branch. The new link is added to the tail of the link list of each node.

The graft! function can also be used to link nodes without need to pass a branch object.

See also graft! and unlink!

Monstera.unlink! — Function

unlink!(branch, node) unlink!(branch => node)

Disconnect a branch from a node.

See also link!

unlink!(node1, node2) -> Branch

Disconnect node1 from node2 and return the branch that was between them.

Monstera.graft! — Function

graft!(tree, graft_node, receiver_node, brlength=NaN)
graft!(tree, graft_node, receiver_branch, newinner=false)

Graft a node (and the subtree to which it is connected) on a receiver branch or node.

Returns nothing, or a new branch-node pair when the newinner option is used (see below).

Grafting on a node

To graft a node on another node is to link the two with a new branch. The length of the new branch can be set with the brlength argument. The function returns the new branch.

Grafting on a branch

Say that you have a branch br1 (1, in the figure below), connected to nodes a and d. To graft a node g directly on branch br1 means that g is placed in between a and d, which results in the creation of a new branch (2, in the figure below). The function returns nothing.

      (a)              (e)                                   (a)                     (e)                    
        \              /                                       \                     /                      
         \            /                                         \                   /                       
          \          /                                           \                 /                        
           \        /            `graft!(tree, g, br1)`           \               /                         
(g)        (a)----(d)         -------------------------->         (a)----(g)----(d)                         
           /   1    \                                             /   1      2    \                         
          /          \                                           /                 \                        
         /            \                                         /                   \                       
        /              \                                       /                     \                      
      (c)              (f)                                   (c)                     (f)

Bipartitions

Monstera.TaxonBipartition — Type

TaxonBipartition

Bit-vector that tags taxa in a list as belonging to one of two disjoint groups.

Beware that many functions assume that bipartitions are normalised so that the first bit is always naught (see Penny & Hendy 1985).

Monstera.bipartitions — Function

bipartitions(tree, singletons=true)

Return a list of taxon bipartitions corresponding to the branches of a tree.

The tree and its nodes must be associated to a taxon set. Singleton bipartitions are excluded by default.

The returned list is of type TaxonBipartitionList. It is possible to retrieve the bipartition corresponding to a branch with the bybranch property.

Monstera.is_singleton — Function

is_singleton(bipartition)

Return true is the bipartition contains a singleton, i.e. it separates a single taxon from all the rest.

In other words, a singleton represents a taxon at a leaf node of a tree. This kind of bipartition is commonly called "trivial", because they provide no additional information about tree topology if one knows that every taxon is a leaf node. But singleton bipartitions are not "trivial" in trees with sampled ancestors.

Monstera.are_compatible — Function

are_compatible(bipartition1, bipartition2)

Check compatibility between two partitions.

Two bipartitions B1 and B2 are compatible if it is logically possible for them to exist in the same tree.

This can be verified with one of the following conditions being true:

B1 = B1 ∪ B2

B1 = B1 ∪ ¬B2

¬B1 = ¬B1 ∪ B2

¬B1 = ¬B1 ∪ ¬B2

Monstera.bipartition_table — Function

bipartition_table(trees, trim=true)

Return a record of the incidences of bipartitions in a collection of trees.

The bipartition table contains a matrix where the rows represent unique bipartitions and the columns represent trees. The entries of the matrix are true when a bipartition occurs in a tree. The matrix can be accessed by indexing the table either by bipartition or by tree index (multiple indices and ranges are not supported yet).

A number of entries are pre-allocated based on a guess of how many unique biparitions may be in the collection. There are removed by default (trim=true).

Properties

(Accessable through the dot-syntax)

bipartitions: The list of unique bipartitions

frequencies: The number of times each bipartition occurs in a tree

n: The number of trees

m: The number of unique bipartitions

Tree distances

Monstera.distance — Function

distance(trees, mode = :ci; <keyword arguments>)
distance(tree1, tree2, mode = :ci; <keyword arguments>)

Return the distance between two trees or a the upper triangle of the matrix of pairwise distances among trees in a collection.

Two distances are available through the mode argument: Clustering Information Distance (:ci) (Smith 2020) and Robinson-Foulds (:rf) (Robinson & Foulds 1981).

It is assumed that all trees belong to a same taxon set and have the same taxa among them, that the trees are unrooted, and that polytomies are hard.

Keyword arguments

singletons=false: if true, singleton bipartitions (corresponding to outer branches) are considered when computing the distances. This deviates from the original definitions ofthe distances, but might be useful for trees with taxa in inner nodes ("sampled ancestors").

For Clustering Information Distance (:ci mode)

diffonly=true: if true, the distance is computed only from the bipartitions that are not shared between the two trees. This makes small distances more accurate and faster to compute.