Base.get — Method

get(taxonset, name)

Get the taxon ID that matches a name in a taxon set, or return 0 if there is no match.

Base.haskey — Method

hasname(taxonset, name)

Check whether a taxon set contains a name.

Base.keys — Method

names(taxonset)

Return the list of names of taxa in a taxon set.

Base.parent — Method

parent(node)
parent(branch => node)

Return the parent of a node.

Throws an error if the node has no parent.

Base.print — Method

Print taxon bipartition grouping taxa by 5

Base.setindex! — Method

Draw on a canvas. Points outside the canvas will be ignored.

Base.size — Method

size(tree)

Return the number of branches and nodes in a tree.

Monstera._clustering_entropy — Function

Return H * n

Monstera._mutual_clustering_information — Function

Return s * n

Monstera.add! — Method

add!(taxonset, names...)

Add name(s) to a taxon set and return its id.

Names already in the taxon set will be ignored (skipped or return 0).

Monstera.are_compatible — Method

are_compatible(bipartition1, bipartition2)

Check compatibility between two partitions.

Two bipartitions B1 and B2 are compatible if it is logically possible for them to exist in the same tree.

This can be verified with one of the following conditions being true:

B1 = B1 ∪ B2

B1 = B1 ∪ ¬B2

¬B1 = ¬B1 ∪ B2

¬B1 = ¬B1 ∪ ¬B2

Monstera.bipartition_table — Function

bipartition_table(trees, trim=true)

Return a record of the incidences of bipartitions in a collection of trees.

The bipartition table contains a matrix where the rows represent unique bipartitions and the columns represent trees. The entries of the matrix are true when a bipartition occurs in a tree. The matrix can be accessed by indexing the table either by bipartition or by tree index (multiple indices and ranges are not supported yet).

A number of entries are pre-allocated based on a guess of how many unique biparitions may be in the collection. There are removed by default (trim=true).

Properties

(Accessable through the dot-syntax)

bipartitions: The list of unique bipartitions

frequencies: The number of times each bipartition occurs in a tree

n: The number of trees

m: The number of unique bipartitions

Monstera.bipartitions — Method

bipartitions(tree, singletons=true)

Return a list of taxon bipartitions corresponding to the branches of a tree.

The tree and its nodes must be associated to a taxon set. Singleton bipartitions are excluded by default.

The returned list is of type TaxonBipartitionList. It is possible to retrieve the bipartition corresponding to a branch with the bybranch property.

Monstera.branch_between — Method

branch_between(node1, node2)

Return the branch that connects node1 to node2.

See also nodes_flanking.

Monstera.branchtype — Function

treetype(x) -> type <: AbstractTree
nodetype(x) -> type <: AbstractNode
branchtype(x) -> type <: AbstractBranch

Return the node, branch, or tree type that matches another type.

This is an auxiliary function used to select matching types in constructors.

Monstera.children — Method

children(branch => node) -> iterator
children(node) -> iterator

Return an iterator of the branches and nodes that are children of a node.

See also neighbours.

Monstera.cladesize — Function

cladesize(branch => node, outer=false)
cladesize(node, outer=false)
cladesize(tree, outer=false)

Return the number of nodes in a clade, optionally counting only the outer nodes.

The clade size of a tree is computed from the tree's root. Unlike the size function, cladesize counts that are directly or indirectly connected to the root.

Monstera.cleanup! — Function

cleanup!(tree)

Remove unconnected nodes and branches and re-index the connected elements.

Monstera.clone — Method

clone(tree)

Create a copy of a tree.

The copied tree has the same topology and properites, but its contents are fully independent. Only the taxon set is shared between the original tree and the clone.

Monstera.clustering_entropy — Function

clustering_entropy(bipartition, n=length(bipartition))

Return the entropy associated to a taxon bipartition as given by the formula from Smith 2020:

$H = (-A/n) * log₂(A/n) - (B/n) * log₂(B/n)$

Where A and B represent the sizes of each bipartition and n represents the total number of taxa.

Monstera.collapse! — Method

collapse!(tree, branch => node)

Eliminate a branch and node in the tree, and reconnect their children to their parent.

In the example below, we collapse branch "1" and node "b".

(c) (d) \ / \ / 3\ 4/ (c) (d) (e) \ / \ | / (b) (e) \ 4| / \ / 3\ |2/ \ / ––––> \|/ 1\ 2/ (a) \ / | (a) | | | | ⃨ | ⃨

Monstera.count_symdiff — Method

count_symdiff(v1, v2)

Count the number of elements that two integer vectors do not have in common.

Zeros are ignored.

Assumptions

Vectors of equal size

Vectors ordered from greatest to smallest

Elements are unique within vectors

Monstera.distance — Function

distance(trees, mode = :ci; <keyword arguments>)
distance(tree1, tree2, mode = :ci; <keyword arguments>)

Return the distance between two trees or a matrix of pairwise distances between trees in a collection.

Two distances are available through the mode argument: Clustering Information Distance (:ci) (Smith 2020) and Robinson-Foulds (:rf) (Robinson & Foulds 1981).

It is assumed that all trees belong to a same taxon set and have the same taxa among them, that the trees are unrooted, and that polytomies are hard.

Keyword arguments

singletons=false: if true, singleton bipartitions (corresponding to outer branches) are considered when computing the distances. This deviates from the original definitions ofthe distances, but might be useful for trees with taxa in inner nodes ("sampled ancestors").

For Clustering Information Distance (:ci mode)

diffonly=true: if true, the distance is computed only from the bipartitions that are not shared between the two trees. This makes small distances more accurate and faster to compute.

Monstera.fast_symdiff — Method

fast_symdiff(v1, v2)

Return the symmetric difference between two vectors of integers.

Zeros are ignored.

This is not general-purpose alternative to symdiff.

Assumptions

Vectors of equal size

Vectors ordered from greatest to smallest

Elements are unique within vectors

Monstera.find_best_assignment — Method

findbestassignment(C, maximize=false) = solution

Solve the two-dimensional assignment problem with a rectangular cost matrix C, scanning row-wise.

Note that cost returned can overflow if using smaller integer types

Example

julia> M=rand(1:100,3,4)
3×4 Matrix{Int64}:
 77  51  42  67
 72  53  47   4
 24  50  77  96

 julia> sol = find_best_assignment(M)
 AssignmentSolution(CartesianIndex.(1:3, [3, 4, 1]), 70)
 
 julia> sum(M[sol])
 70
 
 julia> max_sol = find_best_assignment(M', true)
 AssignmentSolution(CartesianIndex.([1, 2, 4], 1:3), 226)

This code is a port of Matlab code released [1] in the public domain by the US Naval Research Laboratory.

This work is not affliated with or endorsed by the US Naval Research Laboratory.

The algorithm is described in detail in [2].

REFERENCES: [1] D. F. Crouse, "The Tracker Component Library: Free Routines for Rapid Prototyping," IEEE Aerospace and Electronic Systems Magazine, vol. 32, no. 5, pp. 18-27, May. 2017 [2] D. F. Crouse, "Advances in displaying uncertain estimates of multiple targets," in Proceedings of SPIE: Signal Processing, Sensor Fusion, and Target Recognition XXII, vol. 8745, Baltimore, MD, Apr. 2013

Monstera.find_path — Method

find_path(node1, node2)

Find the path from node1 to node2 in a tree.

Returns a tuple with the path as a vector of nodes and a vector of branches. Raises an error if no path is found.

Examples

julia> tree = Newick.parse(RoundaboutTree, "((5,2),((4,1),3));");

# julia> textplot(tree, outerlabels=false)
#      ┌───────○(3)
# ┌────○(2)
# │    └───────○(4)
# ○(1)
# │            ┌───────○(7)
# │    ┌───────○(6)
# └────○(5)    └───────○(8)
#      │
#      └───────○(9)

# Find the path between node 4 and node 6

julia> path_nodes, path_branches = find_path(tree.nodes[4], tree.nodes[6]);

julia> path_nodes
5-element Vector{RoundaboutNode}:
 RoundaboutNode #4 (no taxon) - "2"
 RoundaboutNode #2 (no taxon)
 RoundaboutNode #1 (no taxon)
 RoundaboutNode #5 (no taxon)
 RoundaboutNode #6 (no taxon)

julia> path_branches
4-element Vector{RoundaboutBranch}:
 RoundaboutBranch #3: 2 ○⎯⎯⎯○ 4 (length NaN)
 RoundaboutBranch #1: 1 ○⎯⎯⎯○ 2 (length NaN)
 RoundaboutBranch #4: 1 ○⎯⎯⎯○ 5 (length NaN)
 RoundaboutBranch #5: 5 ○⎯⎯⎯○ 6 (length NaN)

Monstera.find_prev_and_next_branches — Method

Return the next branch and side, and the previous branch and side in a branch cycle

Monstera.get_link_fields — Method

Return the fields of a branch necessary to move on to the next branch on the left or right side.

Monstera.getnode — Method

Return the node that is connected to a given side of the branch

Monstera.getside — Method

Return the side of the branch (left, right or no side) that is connected to the node.

Monstera.graft! — Function

graft!(tree, graft_node, receiver_node, brlength=NaN)
graft!(tree, graft_node, receiver_branch, newinner=false)

Graft a node (and the subtree to which it is connected) on a receiver branch or node.

Returns nothing, or a new branch-node pair when the newinner option is used (see below).

Grafting on a node

To graft a node on another node is to link the two with a new branch. The length of the new branch can be set with the brlength argument. The function returns the new branch.

Grafting on a branch

Say that you have a branch br1 (1, in the figure below), connected to nodes a and d. To graft a node g directly on branch br1 means that g is placed in between a and d, which results in the creation of a new branch (2, in the figure below). The function returns nothing.

      (a)              (e)                                   (a)                     (e)                    
        \              /                                       \                     /                      
         \            /                                         \                   /                       
          \          /                                           \                 /                        
           \        /            `graft!(tree, g, br1)`           \               /                         
(g)        (a)----(d)         -------------------------->         (a)----(g)----(d)                         
           /   1    \                                             /   1      2    \                         
          /          \                                           /                 \                        
         /            \                                         /                   \                       
        /              \                                       /                     \                      
      (c)              (f)                                   (c)                     (f)

Monstera.hasparent — Method

hasparent(node)

Check whether a node has a parent.

Monstera.highight! — Method

Set coordinates in a canvas to appear highlighted (blinking + reversed).

Monstera.is_singleton — Method

is_singleton(bipartition)

Return true is the bipartition contains a singleton, i.e. it separates a single taxon from all the rest.

In other words, a singleton represents a taxon at a leaf node of a tree. This kind of bipartition is commonly called "trivial", because they provide no additional information about tree topology if one knows that every taxon is a leaf node. But singleton bipartitions are not "trivial" in trees with sampled ancestors.

Monstera.isinner — Method

isinner(node) -> Bool

Check whether a node is inner (a.k.a. "internal").

Inner nodes have a degree of 2 or greater.

See also isouter.

Monstera.isouter — Method

isouter(node) -> Bool

Check whether a node is outer (a.k.a. "external", "tip", or "leaf").

Outer nodes always have a degree of 1. This can include the tree root and isolated nodes.

See also isinner.

Monstera.key — Method

key(taxonset, ids...)

Get the names of the taxa with the given ids in the taxon set.

Monstera.link! — Method

link!(node1, branch, node2)

Connect node1 to node2 through a branch.

Linking order is left-to-right, so that node1 is connected on the left side of the branch and node2 is connected on the right side of the branch. The new link is added to the tail of the link list of each node.

The graft! function can also be used to link nodes without need to pass a branch object.

See also graft! and unlink!

Monstera.mci_cost_matrix — Method

Return matrix with s * n value of each bipartition matching.

Monstera.mutual_clustering_information — Method

mutual_clustering_information(bipartition1, bipartition2)

Return the mutual clustering information between two bipartitions.

Monstera.neighbours — Method

neighbours(node) -> iterator

Return an iterator of the branches and nodes that are connected to a given node.

See also children.

Monstera.newbranch! — Method

newnode!(tree, n=1)
newbranch!(tree, n=1)

Create and return n new branches or nodes in a tree.

Monstera.newnode! — Method

newnode!(tree, n=1)
newbranch!(tree, n=1)

Create and return n new branches or nodes in a tree.

Monstera.nodes_flanking — Method

nodes_flanking(branch)

Return a tuple with the nodes connected to a branch.

When nodes are missing from one side of the branch, nothing will take their place in the tuple.

See also branch_between.

Monstera.nodetype — Function

treetype(x) -> type <: AbstractTree
nodetype(x) -> type <: AbstractNode
branchtype(x) -> type <: AbstractBranch

Return the node, branch, or tree type that matches another type.

This is an auxiliary function used to select matching types in constructors.

Monstera.normalise! — Method

normalise!(bipartition)

Normalise a bipartition so that the first bit is always naught (Penny & Hendy 1985)

Monstera.paint_white_background! — Method

Replace non-line break characters with from the left up to the x-th character at height y

Monstera.pretraverse — Method

pretraverse(f, tree)
pretraverse(f, branch => node)
pretraverse(f, node)

Traverse a tree (or subtree) in preorder, calling the function f on each branch-node pair as they are visited.

Returns the return value of the last call of f.

Monstera.prune! — Function

prune!(branch1, node, branch2)

Monstera.randtree — Function

randtree(T, taxonset, rooted=true)
randtree(T, n, rooted=true)

Create a random binary tree of type T with the given taxon set, or the given number n of outer nodes.

If rooted is false, the tree will be unrooted, i.e. it will have a basal trichotomy (this requires at least 3 outer nodes).

Random.seed!(1)
textplot(randtree(RoundaboutTree, 5, false))

# output

┌────○(2)
│
│    ┌───────○(5)
○(1)─○(3)
│    └───────○(6)
│
│    ┌───────○(7)
└────○(4)
     └───────○(8)

Monstera.raw_bipartitions — Method

raw_bipartitions(tree)

Return a the raw (=unnormalised) bipartitions of a tree, singletons included.

Monstera.swapnodes! — Method

swapnodes!(branch1 => node1, branch2 => node2) -> nothing

Exchange nodes between two pairs of nodes and branches.

Monstera.textplot — Method

textplot([io::IO=stdout], tree; <keyword arguments>)
textplot([io::IO=stdout], node; <keyword arguments>)
textplot([io::IO=stdout], branch => node; <keyword arguments>)

Use characters to plot a tree or subtree.

Warning: Branch lengths are not respected!

Arguments

ids=true: Show the ids of the nodes (between parentheses).

outerlabels=true: Show the labels of outer nodes.

innerlabels=false: Show the labels of inner nodes.

highlight=Int[]: Vector of nodes or node IDs that will appear highlighted in the plot.

Examples

newick = "((A,B)H,((C,D,E)I,(F)K)J)G;"
tree = Newick.parse(RoundaboutTree, newick, TaxonSet())
textplot(tree)

# output

     ┌───────●(3)A
┌────○(2)
│    └───────●(4)B
│
○(1)         ┌───────●(7)C
│            │
│    ┌───────○(6)────●(8)D
│    │       │
└────○(5)    └───────●(9)E
     │
     └───────○(10)────●(11)F

Monstera.treetype — Function

treetype(x) -> type <: AbstractTree
nodetype(x) -> type <: AbstractNode
branchtype(x) -> type <: AbstractBranch

Return the node, branch, or tree type that matches another type.

This is an auxiliary function used to select matching types in constructors.

Monstera.two_way_setdiff — Method

two_way_setdiff(v1, v2, result1, result2)

Compute setdiff(v1, v2) and store it on result1, and compute setdiff(v2, v1) and store it in result2

Returns the indices of the last element stored in each result vector. Result vectors must be of the same length as input vectors. Zeros are ignored.

This is not general-purpose alternative to symdiff.

Assumptions

Vectors of equal size

Vectors ordered from greatest to smallest

Elements are unique within vectors

Monstera.ungraft! — Method

ungraft!(tree, branch => node)

Separate a node with a branch from the rest of the tree without leaving other nodes disconnected.

Monstera.unlink! — Method

unlink!(branch, node) unlink!(branch => node)

Disconnect a branch from a node.

See also link!

Monstera.unlink! — Method

unlink!(node1, node2) -> Branch

Disconnect node1 from node2 and return the branch that was between them.

Monstera.AbstractBranch — Type

The AbstractBranch type is the supertype of branch types. Subtypes of AbstractBranch must implement the public properties listed below.

Public properties

id: integer number that uniquely identifies a node in a tree. Not user-settable.

length: branch length as a Float64. Branches with NA values are considered to

have no length. User-settable.

Monstera.AbstractNode — Type

The AbstractNode type is the supertype of node types. Subtypes of AbstractNode must implement the public properties listed below.

Public properties

id: integer number that uniquely identifies a node in a tree. Not user-settable.

taxon: integer number that identifies the taxon associated to a node. User-settable.

label: string to display for this node when printing, plotting, etc. User-settable.

Monstera.AssignmentSolution — Type

Stores the solution of an assigment problem

Can be directly used to index the appropriate values from the cost matrix

Monstera.CharCanvas — Type

CharCanvas

A canvas for drawing text plots upon.

Drawing coordinates are in "X,Y" format with the origin at the top right corner.

Canvases are created filled with , and the first occurence of in a line causes printing to skip the rest of the line and move on to the next. The paint_white_background function can be used to change ' ' to ' ' in selected areas. This awkward system exists to prevent printing excessive right-padding whitespaces, which makes tree plots more prone to deforming when resizing the text area.

Monstera.TaxonBipartition — Type

TaxonBipartition

Bit-vector that tags taxa in a list as belonging to one of two disjoint groups.

Beware that many functions assume that bipartitions are normalised so that the first bit is always naught (see Penny & Hendy 1985).

Monstera.TaxonSet — Type

TaxonSet([taxon_names])
TaxonSet(n::Int)

An ordered collection of taxa with unique names.

TaxonSets can be created empty, or from a vector of name strings, or by giving a number n, which creates taxa successively named "1" through "n".

Monstera.Newick.parse — Method

Newick.parse(T, string, settings)

Parse a string in Newick format as a tree of type T.

Monstera.Newick.parse_branch! — Method

After reading ':'

Monstera.Newick.parse_comment — Method

After reading '['. Doesn't work with arrays yet!

Monstera.Newick.parse_quoted — Method

Parse a quoted string after Base.reading the first quote (qchar to specify ''' vs '"'), with escape character escchar

Monstera.Newick.read — Method

Newick.read(file, sink)
Newick.read(io, sink)

Read trees from a file in Newick format.

The sink argument specifies the concrete type that will be used to represent the trees. It can correspond to an AbstractTree subtype or a vector thereof. If a tree type is given, only the first Newick string in the file will be read. If a vector of a tree type is given, all the Newick strings in the file will be read.

Monstera.Newick.string — Method

Newick.string(tree, <keyword arguments or writer settings object>)
Newick.string(node, <keyword arguments or writer settings object>)
Newick.string(branch => node, <keyword arguments or writer settings object>)

Return a string with the Newick representation of a tree.

The string format can be optionally customised by passing a WriterSettings object, or with the keyword arguments below.

See also Newick.read, Newick.WriterSettings

Arguments

brlength::Bool=true: whether to write branch lengths

label_quote::Bool=false: wheteher to write double quote marks around node labels

brlength_digits::Int=308: branch lengths will be rounded to this number

skip_nanlength::Bool=false: whether to write the length of a branch when it is NaN

Monstera.Newick.write — Method

Newick.write(<file name or io>, tree, <keyword arguments or writer settings object>)
Newick.write(<file name or io>, node, <keyword arguments or writer settings object>)
Newick.write(<file name or io>, branch => node, <keyword arguments or writer settings object>)

Write the Newick representation of a tree to an IOBuffer or a file.

The Newick string format can be optionally customised by passing a NewickWriterSettings object, or with the keyword arguments below.

See also Newick.string, Newick.WriterSettings

Arguments

brlength::Bool=true: whether to write branch lengths

label_quote::Bool=false: wheteher to write double quote marks around node labels

brlength_digits::Int=308: branch lengths will be rounded to thisnumber

skip_nanlength::Bool=false: whether to write the length of a branch when it is NaN

Writing to file

append=false: whether to append to the file

lock=true: lock operations for safe multi-threaded access.

Monstera.Newick.ReaderSettings — Type

Newick.ReaderSettings(taxonset = TaxonSet(); taxafrom=:outer)

Settings for reading Newick strings.

Node labels can be interpreted as taxon names. Taxon names will be mapped to the corresponding taxon in the taxon set, or added to the taxon set if they were not already present.

The taxafrom keyword argument is used to customise which node labels will be interpeted as taxon names, with any of the following values:

:outer (default) only outer nodes

:inner only inner nodes

:all all nodes

:none no nodes

Monstera.Newick.WriterSettings — Type

Newick.WriterSettings(args)

Settings for writing Newick strings.

Arguments

brlength::Bool=true: whether to write branch lengths

label_quote::Bool=false: wheteher to write double quote marks around node labels

brlength_digits::Int=308: branch lengths will be rounded to this number of digits

skip_nanlength::Bool=true: whether to write the length of a branch when it is NaN